hopr_protocol_start

Enum StartProtocol

pub enum StartProtocol<I, T, C> {
    StartSession(StartInitiation<T, C>),
    SessionEstablished(StartEstablished<I>),
    SessionError(StartErrorType),
    KeepAlive(KeepAliveMessage<I>),
}

Lists all messages of the Start protocol for a session establishment.

Generic parameters

I is the session identifier.

T is the session target.

C are session capabilities.

Diagram of the protocol

sequenceDiagram Entry->>Exit: SessionInitiation (Challenge) alt If Exit can accept a new session Note right of Exit: SessionID [Pseudonym, Tag] Exit->>Entry: SessionEstablished (Challenge, SessionID_Entry) Note left of Entry: SessionID [Pseudonym, Tag] Entry->>Exit: KeepAlive (SessionID) Note over Entry,Exit: Data else If Exit cannot accept a new session Exit->>Entry: SessionError (Challenge, Reason) end opt If initiation attempt times out Note left of Entry: Failure end

Variants

StartSession(StartInitiation<T, C>)

Request to initiate a new session.

SessionEstablished(StartEstablished<I>)

Confirmation that a new session has been established by the counterparty.

SessionError(StartErrorType)

Counterparty could not establish a new session due to an error.

KeepAlive(KeepAliveMessage<I>)

A ping message to keep the session alive.

Implementations

impl<I, T, C> StartProtocol<I, T, C>

pub const START_PROTOCOL_MESSAGE_TAG: Tag

Fixed [Tag] of every protocol message.

pub const START_PROTOCOL_VERSION: u8 = 0x02

Current version of the Start protocol.

impl<I, T, C> StartProtocol<I, T, C>
where I: Serialize + for<'de> Deserialize<'de>, T: Serialize + for<'de> Deserialize<'de>, C: Into<u8> + TryFrom<u8>,

pub fn encode(self) -> Result<(Tag, Box<[u8]>)>

Tries to encode the message into binary format and [Tag]

pub fn decode(tag: Tag, data: &[u8]) -> Result<Self>

Tries to decode the message from the binary representation and [Tag].

The tag must be currently START_PROTOCOL_MESSAGE_TAG and version START_PROTOCOL_VERSION.