Crate hopr_protocol_start

This crate defines the Start sub-protocol used for HOPR Session initiation and management.

The Start protocol is used to establish Session as described in HOPR RFC-0012. and is implemented via the StartProtocol enum.

The protocol is defined via generic arguments I (for Session ID), T (for Session Target) and C (for Session capabilities).

Per RFC-0012, the types I and T are serialized/deserialized to the CBOR binary format (see RFC7049) and therefore must implement serde::Serialize + serde::Deserialize. The capability type C must be expressible as a single unsigned byte.

See StartProtocol docs for the protocol diagram.

Modules

errors

Contains errors raised by the Start protocol.

Structs

KeepAliveMessage

Keep-alive message for a Session with the identifier T.

StartErrorType

Error message in the Start protocol.

StartEstablished

Message of the Start protocol that confirms the establishment of a session.

StartInitiation

The session initiation message of the Start protocol.

Enums

StartErrorReason

Lists all Start protocol error reasons.

StartProtocol

Lists all messages of the Start protocol for a session establishment.

StartProtocolDiscriminants

Type Aliases

StartChallenge

Challenge that identifies a Start initiation protocol message.